Move encoding from editor to taylor, add UI to test native functions
