+++ web/src/components/TaylorInterpreter.js
@@ -1,8 +1,9 @@
 import React, { Component } from 'react';
-import { View, Item, Input, Text, Button, Icon } from 'native-base';
+import { View, Item, Input, Text, Button, Icon, Picker } from 'native-base';
 import { getProvider } from '../utils/web3.js';
 import { getStoredTypes, saveType, executeType } from '../utils/interpreter.js';
 import { storeAddress, getAddress } from '../utils/taylor.js';
+import * as taylorUtils from 'taylor/taylor/interpreter.js';
 
 class TaylorInterpreter extends Component {
   constructor(props) {
@@ -15,6 +16,7 @@ class TaylorInterpreter extends Component {
       types: [],
       currentType: {},
       currentUserInputs: null,
+      currentSignature: null,
     }
 
     this.onChangeAddress = this.onChangeAddress.bind(this);
@@ -23,6 +25,7 @@ class TaylorInterpreter extends Component {
     this.onTypeSave = this.onTypeSave.bind(this);
     this.changeUserInputs = this.changeUserInputs.bind(this);
     this.onExecute = this.onExecute.bind(this);
+    this.changeCurrentSig2 = this.changeCurrentSig2.bind(this);
 
     this.setWeb3();
   }
@@ -69,15 +72,19 @@ class TaylorInterpreter extends Component {
     this.setState({ currentType });
   }
 
+  changeCurrentSig2(currentSignature) {
+    this.setState({ currentSignature });
+  }
+
   changeUserInputs(currentUserInputs) {
     this.setState({ currentUserInputs });
   }
 
   async onExecute() {
-    const { currentUserInputs, currentType, provider, addressData } = this.state;
+    const { currentUserInputs, currentType, provider, addressData, currentSignature } = this.state;
     let args = (currentUserInputs || '').split(',');
-    args = encodeInput(args);
-    args = strip0x(currentType.signature) + args;
+    args = taylorUtils.encodeInput(args);
+    args = taylorUtils.strip0x(currentSignature || currentType.signature) + args;
 
     const executeAnswer = await executeType(provider)(addressData.address)(args);
     this.setState({ executeAnswer });
@@ -89,9 +96,9 @@ class TaylorInterpreter extends Component {
     let encodedData;
     if (result.result) {
       const { steps, tr } = result.result[0];
-      const {hadc_inputs, user_inputs, prog_steps} = buildInputs(steps, tr);
+      const {hadc_inputs, user_inputs, prog_steps} = taylorUtils.buildInputs(steps, tr);
 
-      encodedData = buildType(currentType.name, currentType.signature, {hadc_inputs, user_inputs, prog_steps});
+      encodedData = taylorUtils.buildType(currentType.name, currentType.signature, {hadc_inputs, user_inputs, prog_steps});
 
       saveType(signer)(addressData.address)(encodedData.encoded);
     }
@@ -100,186 +107,79 @@ class TaylorInterpreter extends Component {
   render() {
     const {styles} = this.props;
     const { types, addressData, executeAnswer } = this.state;
-    console.log('types', types);
+    const { result, errors } = this.props.currentGraph || {};
+
+    const tfunctions = result && result.result ? result.result[0].tr : {};
+
     return (
-      <View style={{ ...styles, flex: 1 }}>
-        <Item style={{ width: styles.width }}>
-          <Input
-            value={ addressData ? addressData.address : '' }
-            style={{ color: 'white' }}
-            placeholder='interpreter address'
-            onChangeText={this.onChangeAddress}
-          />
-        </Item>
-        <Text style={{ color: 'grey' }}>{JSON.stringify(types)}</Text>
+      <View style={{ ...styles}}>
+        <View style={{ ...styles, flex: 1 }}>
+          <Item style={{ width: styles.width }}>
+            <Input
+              style={{ color: 'white' }}
+              placeholder='name'
+              label='name'
+              onChangeText={this.changeCurrentName}
+            />
+          </Item>
+          <Item style={{ width: styles.width }}>
+            <Input
+              style={{ color: 'white' }}
+              placeholder='signature'
+              label='signature'
+              onChangeText={this.changeCurrentSig}
+            />
+          </Item>
+          <Button small light onClick={this.onTypeSave}>
+            <Icon name='save' />
+          </Button>
+        </View>
+        <View style={{ ...styles, flex: 1 }}>
+          <Item style={{ width: styles.width }}>
+            <Input
+              value={ addressData ? addressData.address : '' }
+              style={{ color: 'white' }}
+              placeholder='interpreter address'
+              onChangeText={this.onChangeAddress}
+            />
+          </Item>
+          <Text style={{ color: 'grey' }}>{JSON.stringify(types)}</Text>
 
-        <Item style={{ width: styles.width }}>
-          <Input
-            style={{ color: 'white' }}
-            placeholder='name'
-            label='name'
-            onChangeText={this.changeCurrentName}
-          />
-        </Item>
-        <Item style={{ width: styles.width }}>
-          <Input
-            style={{ color: 'white' }}
-            placeholder='signature'
-            label='signature'
-            onChangeText={this.changeCurrentSig}
-          />
-        </Item>
-        <Button small light onClick={this.onTypeSave}>
-          <Icon name='save' />
-        </Button>
+          <Item picker>
+            <Picker
+              mode="dropdown"
+              style={{ width: styles.width, backgroundColor: 'rgb(169, 169, 169)' }}
+              placeholder="Select your SIM"
+              placeholderStyle={{ color: "#bfc6ea" }}
+              placeholderIconColor="#007aff"
+              selectedValue={this.state.currentSignature}
+              onValueChange={this.changeCurrentSig2}
+            >
+              <Picker.Item label="select signature" value={null} />
+              {
+                Object.keys(tfunctions).map(name => {
+                  return <Picker.Item label={name} value={tfunctions[name].sig} />
+                })
+              }
+            </Picker>
+          </Item>
 
-        <Item style={{ width: styles.width }}>
-          <Input
-            style={{ color: 'white' }}
-            placeholder='user input'
-            label='user input'
-            onChangeText={this.changeUserInputs}
-          />
-        </Item>
-        <Button small light onClick={this.onExecute}>
-          <Icon name='play' />
-        </Button>
-        <Text style={{ color: 'grey' }}>{JSON.stringify(executeAnswer)}</Text>
+          <Item style={{ width: styles.width }}>
+            <Input
+              style={{ color: 'white' }}
+              placeholder='user input'
+              label='user input'
+              onChangeText={this.changeUserInputs}
+            />
+          </Item>
+          <Button small light onClick={this.onExecute}>
+            <Icon name='play' />
+          </Button>
+          <Text style={{ color: 'grey' }}>{JSON.stringify(executeAnswer)}</Text>
+        </View>
       </View>
     );
   }
 }
 
 export default TaylorInterpreter;
-
-
-
-function buildInputs(steps, tr) {
-  const user_inputs = [];
-  const hadc_inputs = [];
-  const prog_steps = [];
-
-  console.log('steps', steps)
-  steps.forEach(step => {
-    console.log('step', step)
-    const fname = step[0]
-    const args = step[1];
-    let arglen = args ? args.length : 0;
-    console.log('--arglen', arglen)
-    const prog_step = {
-      typeid: tr[fname] ? tr[fname].sig : '',
-      inputIndexes: [],
-    }
-
-    const user_inputs_step = [];
-    const hardc_inputs_step = [];
-    for ( let i = 0; i < arglen; i ++ ) {
-      if (typeof args[i] === 'string' && args[i].substring(0, 2) === 'x_' && args[i].length >= 3) {
-        user_inputs_step.push(args[i]);
-      } else if (!isNaN(args[i])) {
-        hardc_inputs_step.push(args[i]);
-      }
-    }
-
-    console.log('user_inputs_step', user_inputs_step)
-    console.log('hardc_inputs_step', hardc_inputs_step)
-
-    for ( let i = 0; i < arglen; i ++ ) {
-      console.log('args[i]', args[i]);
-
-      console.log('user_inputs', user_inputs)
-      console.log('hadc_inputs', hadc_inputs)
-
-      if (typeof args[i] === 'string' && args[i].substring(0, 2) === 'x_' && args[i].length >= 3) {
-        prog_step.inputIndexes.push(user_inputs.length);
-        user_inputs.push(args[i]);
-      }
-      else if (isNaN(args[i])) {
-        prog_step.inputIndexes.push(
-          user_inputs_step.length + user_inputs.length
-          + hardc_inputs_step.length + hadc_inputs.length
-        );
-      } else {
-        prog_step.inputIndexes.push(
-          Math.max(user_inputs_step.length, user_inputs.length)
-          + hadc_inputs.length
-        );
-        hadc_inputs.push(args[i]);
-
-      }
-      console.log('999999999', JSON.stringify(hadc_inputs), JSON.stringify(user_inputs))
-    };
-
-    if (tr[fname]) {
-      prog_steps.push(prog_step);
-    }
-  });
-
-  return {user_inputs, hadc_inputs, prog_steps};
-}
-
-function encode(value, size) {
-  return value.toString(16).padStart(size*2, '0');
-}
-
-const isInt = value => parseInt(value) === value;
-const isUint = value => isInt(value) && (value >= 0);
-const isFloat = value => parseFloat(value) === value;
-
-const isIntDyn = value => parseInt(value) == value;
-const isUintDyn = value => isIntDyn(value) && (value >= 0);
-const isFloatDyn = value => parseFloat(value) == value;
-
-
-function encodeDynamic(value) {
-  if (isUint(value)) return '11000004' + encode(value, 4);
-  if (isInt(value)) return '12000004' + encode(value, 4);
-
-  if (isUintDyn(value)) return '11000004' + encode(value, 4);
-  if (isIntDyn(value)) return '12000004' + encode(value, 4);
-
-  throw new Error('unsupported type');
-}
-
-function strip0x(value) {
-  if (value.substring(0, 2) === '0x') {
-    return value.substring(2);
-  }
-}
-
-function encodeInput (args) {
-  args = args.map(encodeDynamic);
-  console.log('args', args);
-  args = 'ee'
-    + encode(args.length, 3)
-    + args.map((arg, i) => {
-        return encode(
-          args.slice(0, i+1).reduce((sum, val) => sum + val.length/2, 0),
-          4,
-        )
-      }).join('')
-    + args.join('')
-  return args;
-}
-
-function buildType(name, signature, {user_inputs, hadc_inputs, prog_steps}) {
-  let dtype = '';
-  let steps = '';
-  let hcinputs = encodeInput(hadc_inputs);
-  hcinputs = encode(hcinputs.length / 2, 4) + hcinputs;
-
-  dtype += strip0x(signature) + encode(prog_steps.length, 1);
-  dtype = encode(dtype.length / 2, 4) + dtype
-
-  prog_steps.forEach(step => {
-    steps += step.typeid;
-    steps += encode(step.inputIndexes.length, 4)
-    step.inputIndexes.forEach(ind => {
-      steps += encode(ind, 1)
-    });
-  })
-
-  steps = encode(steps.length / 2, 4) + steps
-
-  return { dtype, steps, hcinputs, encoded: dtype + hcinputs + steps }
-}
